server/requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
huggingface-hub==0.23.4
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.4
python-dotenv==1.0.1
scikit-learn==1.5.0
scipy==1.13.1
streamlit==1.36.0
streamlit-option-menu==0.3.13
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.3
uvicorn==0.30.1
pyyaml
streamlit_lottie
streamlit_js_eval
